first commit. Interface for sinusoidal pattern profilometry has been added along with stubs functions in the implementation class. The new interface inherits from StructuredLightPattern
Added markers to sinusoidal patterns

computePhaseMap for Fourier transform profilometry

computePhaseMap for Fourier transform profilometry

added phase map computation for PSP. Changed the maskDftRegion to frequencyFiltering. It uses regions of interest and can filter symmetrically. Also added computeShadowMask and computeDataModulationTerm

changed formatting in structured light module. First commit for the phase unwrapping module. So far, pixel reliabilities are computed and edges are sorted in a histogram.

Fixed an error in edges sorting. Added the unwrap histogram method. It computes the number of 2*Pi that has to be added to each pixel to unwrap the phase map

added an example for phase unwrapping and a test that unwraps a simple phase map

Added documentation draft and a small example that can generate sinusoidal patterns

removed unnecessary include

added a few comments in sinusoidalpattern.cpp and histogramphaseunwrapping.cpp. Removed some redudancy about mask in the reliability computation. Changed formatting